Refactor Start + Encoding fix
